feat: refactored backend.py
